requirements.txt
einops==0.3.2
h5py==3.10.0
hickle==5.0.2
hydra-core==1.1.1
imageio==2.33.1
matplotlib>=3.7
moviepy==1.0.3
numpy>=1.24
opencv-python
pandas==2.0.3
pillow==10.2.0
protobuf==3.20.*
psutil==5.8.0
Requests==2.31.0
tensorboard
torch==1.12.1
torchvision==0.13.1
tqdm==4.62.3
